LayoutTests/fast/forms/file/display-none-upload-button.html

<!DOCTYPE HTML>
<style>
input {
    border: 5px solid salmon;
}
input::-webkit-file-upload-button {
    display: none;
}
</style>
Tests that a display:none upload button works.
<input type="file">
<script src="../../js/resources/js-test-pre.js"></script>
<script>
// FIXME: The text inside the file input should give the input a contentHeight
// even when there's no upload button.
// https://bugs.webkit.org/show_bug.cgi?id=109104
shouldBeTrue("document.querySelector('input').offsetHeight > 10");
</script>
<script src="../../js/resources/js-test-post.js"></script>
