third_party/blink/web_tests/platform/generic/fast/forms/file/display-none-upload-button-expected.txt

PASS document.querySelector('input').offsetHeight > 10 is true
PASS successfullyParsed is true

TEST COMPLETE

Tests that a display:none upload button works. 
